+++ docker/generate.py
@@ -264,6 +264,7 @@ def install_build_packages(manager):
             libcurl-devel \\
             # used by Vitis AI
             glog-devel \\
+            gflags-devel \\
             # used by Drogon and Vitis AI
             openssl-devel \\
             # used by TF/PT ZenDNN building
@@ -527,7 +528,7 @@ def install_dev_packages(manager: PackageManager, core):
 
 
 migraphx_apt_repo = 'echo "deb [arch=amd64 trusted=yes] http://repo.radeon.com/rocm/apt/5.0/ ubuntu main" > /etc/apt/sources.list.d/rocm.list'
-migraphx_yum_repo = '[ROCm]\\nname=ROCm\\nbaseurl=https://repo.radeon.com/rocm/yum/5.0/\\nenabled=1\\ngpgcheck=1\\ngpgkey=https://repo.radeon.com/rocm/rocm.gpg.key" > /etc/yum.repos.d/rocm.repo'
+migraphx_yum_repo = '"[ROCm]\\nname=ROCm\\nbaseurl=https://repo.radeon.com/rocm/yum/5.0/\\nenabled=1\\ngpgcheck=1\\ngpgkey=https://repo.radeon.com/rocm/rocm.gpg.key" > /etc/yum.repos.d/rocm.repo'
 
 
 def build_migraphx(manager: PackageManager):