+++ src/proteus/bindings/python/clients/grpc.cpp
@@ -45,12 +45,12 @@ void wrapGrpcClient(py::module_ &m) {
     .def("modelMetadata", &GrpcClient::modelMetadata, py::arg("model"),
          DOCS(GrpcClient, modelMetadata))
     .def("modelLoad", &GrpcClient::modelLoad, py::arg("model"),
-         py::arg("parameters") = proteus::RequestParameters(),
+         py::arg("parameters") = static_cast<RequestParameters *>(nullptr),
          DOCS(GrpcClient, modelLoad))
     .def("modelUnload", &GrpcClient::modelUnload, py::arg("model"),
          DOCS(GrpcClient, modelUnload))
     .def("workerLoad", &GrpcClient::workerLoad, py::arg("model"),
-         py::arg("parameters") = proteus::RequestParameters(),
+         py::arg("parameters") = static_cast<RequestParameters *>(nullptr),
          DOCS(GrpcClient, workerLoad))
     .def("workerUnload", &GrpcClient::workerUnload, py::arg("model"),
          DOCS(GrpcClient, workerUnload))