+++ src/proteus/bindings/python/clients/websocket.cpp
@@ -52,10 +52,15 @@ void wrapWebSocketClient(py::module_ &m) {
     .def("modelMetadata", &WebSocketClient::modelMetadata, py::arg("model"),
          DOCS(WebSocketClient, modelMetadata))
     .def("modelLoad", &WebSocketClient::modelLoad, py::arg("model"),
-         py::arg("parameters") = proteus::RequestParameters(),
+         py::arg("parameters") = static_cast<RequestParameters *>(nullptr),
          DOCS(WebSocketClient, modelLoad))
     .def("modelUnload", &WebSocketClient::modelUnload, py::arg("model"),
          DOCS(WebSocketClient, modelUnload))
+    .def("workerLoad", &WebSocketClient::workerLoad, py::arg("model"),
+         py::arg("parameters") = static_cast<RequestParameters *>(nullptr),
+         DOCS(WebSocketClient, workerLoad))
+    .def("workerUnload", &WebSocketClient::workerUnload, py::arg("model"),
+         DOCS(WebSocketClient, workerUnload))
     .def("modelInfer", &WebSocketClient::modelInfer, py::arg("model"),
          py::arg("request"), DOCS(WebSocketClient, modelInfer))
     // cannot wrap future directly in Python